Move Design System 'dummy content' into 'helpers' folder
