+++ Source/DesignSystem/atoms/AlertDialog/helpers.ts
@@ -15,10 +15,3 @@ Most alerts do not need titles. They summarize a decision in a sentence or two b
 statement related to the action buttons.
 Alert dialogs must always contain an action item such as "agree", "accept", or "continue" and a dismissive action such as "cancel", 
 "disagree" or "deny".`;
-
-export const listArray = [
-    'List item 1', 'List item 2', 'List item 3', 'List item 4', 'List item 5', 'List item 6', 'List item 7',
-    'List item 8', 'List item 9', 'List item 10', 'List item 11', 'List item 12', 'List item 13', 'List item 14',
-    'List item 15', 'List item 16', 'List item 17', 'List item 18', 'List item 19', 'List item 20', 'List item 21',
-    'List item 22', 'List item 23', 'List item 24', 'List item 25', 'List item 26', 'List item 27', 'List item 28',
-];
