+++ Source/DesignSystem/helpers/DummyContents/DummyArray.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+export const dummyArray = [
+    'List item 1', 'List item 2', 'List item 3', 'List item 4', 'List item 5', 'List item 6', 'List item 7',
+    'List item 8', 'List item 9', 'List item 10', 'List item 11', 'List item 12', 'List item 13', 'List item 14',
+    'List item 15', 'List item 16', 'List item 17', 'List item 18', 'List item 19', 'List item 20', 'List item 21',
+    'List item 22', 'List item 23', 'List item 24', 'List item 25', 'List item 26', 'List item 27', 'List item 28',
+];
