+++ Source/DesignSystem/helpers/DummyContents/index.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+export * from './DummyAccordionsContent';
+export * from './DummyArray';
+export * from './DummyText';
