+++ Source/DesignSystem/helpers/ReactRouter/index.tsx
@@ -1,0 +1,5 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+export * from './ReactRouter';
+export * from './DummyLinks';
