--- Source/DesignSystem/helpers/RouterLinkListItem.stories.tsx
@@ -1,26 +1,0 @@
-// Copyright (c) Dolittle. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-// import React from 'react';
-
-// import { componentStories, RouterLinkListItem } from '@dolittle/design-system';
-
-// const { metadata, createStory } = componentStories(RouterLinkListItem);
-
-// export default metadata;
-
-// function Content() {
-//     const location = useLocation();
-
-//     return (
-//         <Typography variant='body2' sx={{ pb: 2 }} color="text.secondary">
-//             Current route: {location.pathname}
-//         </Typography>
-//     );
-// }
-
-{/* <Routes>
-    <Route path="*" element={<Content />} />
-</Routes> */}
-
-//export const Default = createStory();
